reduce prezistion from double to float

+++ Firmware/Drivers/Interfaces/position/position.h
@@ -55,7 +55,7 @@ typedef struct {
 typedef struct{
 	POS_Latitude_Flag direction;
 	uint16_t degree;
-	double minute;
+	float minute;
 } POS_Latitude;
 
 /**
@@ -65,7 +65,7 @@ typedef struct{
 typedef struct{
 	POS_Longitude_Flag direction;
 	uint16_t degree;
-	double minute;
+	float minute;
 } POS_Longitude;
 
 /**
